Todos los Metodos funcionando correctamente y terminados

+++ src/main/java/co/edu/umanizales/tads/model/ListSE.java
@@ -1,21 +1,16 @@
 package co.edu.umanizales.tads.model;
 
-import co.edu.umanizales.tads.controller.dto.*;
-import co.edu.umanizales.tads.service.ListSEService;
+import co.edu.umanizales.tads.controller.dto.ReportKidsLocationDTO;
 import lombok.Data;
 import lombok.Getter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Optional;
+
+import static org.apache.logging.log4j.ThreadContext.isEmpty;
+
 
 @Data
 @Getter
@@ -57,6 +52,7 @@ public class ListSE {
             lastNode.orElseThrow(NullPointerException::new).setNext(newNode);
         } catch (NullPointerException e) {
             head = new Node(kid);
+
         }
         size++;
     }
@@ -78,10 +74,12 @@ public class ListSE {
 
     public void addToStart(Kid kid) {
         Node newNode = new Node(kid);
-        try {
+        if(head == null){
+            head=newNode;
+        }
+        else{
             newNode.setNext(head);
-        } catch (NullPointerException e) {
-            head = newNode;
+            head= newNode;
         }
         size++;
     }
@@ -240,7 +238,6 @@ public class ListSE {
             System.out.println("La lista está vacía.");
         }
     }
-
     public void gainPosition(String id, int position, ListSE listSE) {
         try {
             Node temp = this.head;
@@ -258,39 +255,56 @@ public class ListSE {
             int newPosition = position - count;
 
             if (newPosition < 1) {
-                throw new IllegalArgumentException("La nueva posición debe ser mayor o igual a 1.");
+                newPosition = 1; // Si la nueva posición es menor que 1, lo asignamos a 1 para insertar al principio
+            } else if (newPosition > size + 1) {
+                newPosition = size + 1; // Si la nueva posición es mayor que el tamaño de la lista más 1, lo asignamos al final
             }
 
             Kid listCopy = temp.getData();
-            this.deleteByIdentification(id);
+            listSE.deleteByIdentification(temp.getData().getIdentification());
             listSE.addByPosition(listCopy, newPosition);
         } catch (NullPointerException e) {
-            System.out.println("La lista está vacía.");
+            throw new IllegalStateException("La lista está vacía.");
         } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
+            throw new IllegalArgumentException(e.getMessage());
         }
     }
 
 
-    public void backPosition(String id, int position, ListSE listSE){
+
+
+    public void backPosition(String id, int position, ListSE listSE) {
         try {
             Node temp = this.head;
             int count = 1;
 
-            while (temp != null && ! temp.getData().getIdentification().equals(id)){
+            while (temp != null && !temp.getData().getIdentification().equals(id)) {
                 temp = temp.getNext();
-                count ++;
+                count++;
             }
-            int newPosition = position+count-1;
+
+            if (temp == null) {
+                throw new IllegalArgumentException("No se encontró un niño con la identificación especificada.");
+            }
+
+            int newPosition = position - count;
+
+            if (newPosition < 1) {
+                newPosition = 1; // Si la nueva posición es menor que 1, lo asignamos a 1 para insertar al principio
+            } else if (newPosition > size + 1) {
+                newPosition = size + 1; // Si la nueva posición es mayor que el tamaño de la lista más 1, lo asignamos al final
+            }
+
             Kid listCopy = temp.getData();
             listSE.deleteByIdentification(temp.getData().getIdentification());
-            listSE.addByPosition(listCopy , newPosition);
+            listSE.addByPosition(listCopy, newPosition);
         } catch (NullPointerException e) {
-            System.out.println("La lista está vacía o el nodo con la identificación especificada no existe.");
-        } catch (IndexOutOfBoundsException e) {
-            System.out.println("La posición especificada está fuera de los límites de la lista.");
+            throw new IllegalStateException("La lista está vacía.");
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException(e.getMessage());
         }
     }
+
 
     public void orderBoysToStart() {
         try {
@@ -357,14 +371,17 @@ public class ListSE {
 
     public int getCountKidsByLocation(String code) {
         int count = 0;
-        try {
-            Node temp = this.head;
-            while (temp != null && !temp.getData().getLocation().getCode().equals(code)) {
-                temp = temp.getNext();
+        Node temp = head;
+        while (temp != null) {
+            try {
+                String cityCode = temp.getData().getLocation().getCode().substring(0, 8);
+                if (cityCode.equals(code)) {
+                    count++;
+                }
+            } catch (StringIndexOutOfBoundsException e) {
+
             }
-            count = temp != null ? 1 : 0;
-        } catch (NullPointerException e) {
-            System.out.println("Error: Argumento nulo pasado al método.");
+            temp = temp.getNext();
         }
         return count;
     }
@@ -405,24 +422,32 @@ public class ListSE {
             System.out.println("Error: Argumento nulo pasado al método.");
         }
     }
-
-    public void getReportKidsByAgeByGender(byte age, ReportAgeQuantityKidsDTO report) {
-        try {
-            if (this.head == null) {
-                throw new NullPointerException("Lista vacía");
-            }
-
-            Node temp = this.head;
-            while (temp != null) {
-                if (temp.getData().getAge() == age) {
-                    report.updateQuantity(temp.getData().getAge(), temp.getData().getGender());
-                }
-                temp = temp.getNext();
-            }
-
-        } catch (NullPointerException e) {
-            System.out.println("Error: " + e.getMessage());
+    public int informRangeByAge(int first, int last) throws IllegalArgumentException {
+        if (first < 0 || last < 0 || first > last) {
+            throw new IllegalArgumentException("Invalid age range");
         }
+        Node temp = head;
+        int count = 0;
+        while (temp != null){
+            if (temp.getData().getAge() >= first && temp.getData().getAge() <= last){
+                count ++;
+            }
+            temp = temp.getNext();
+        }
+        return count;
+    }
+
+    public void changeToRound() {
+        if (head == null ) {
+            return;
+        }
+
+        Node temp = head;
+        while (temp.getNext() != null) {
+            temp = temp.getNext();
+        }
+
+        temp.setNext(head);
     }
